Move _grid_from_X related tests to test_ice.py as they are now in _ice.py
